requirements.txt
pdfplumber>=0.10.0
google-generativeai>=0.3.0
fastapi
uvicorn
pydantic
python-multipart
google-generativeai
pdfplumber
python-dotenv>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0

# Gmail API
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Machine Learning (LayoutLMv3)
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
seqeval>=1.2.2
scikit-learn>=1.0.0
tensorboard
easyocr
